+++ LiTsap/app/src/main/java/studio/honidot/litsap/source/local/LiTsapLocalDataSource.kt
@@ -16,7 +16,7 @@ class LiTsapLocalDataSource(val context: Context) : LiTsapDataSource {
     }
 
 
-    override suspend fun getHistory(taskIdList: List<String>,passNday:Long): Result<List<History>> {
+    override suspend fun getHistory(taskIdList: List<String>,passNday:Int): Result<List<History>> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
